requirements-dev.txt
eth-ape>=0.4.0,<0.5.0
black==22.6.0
pytest
ape-etherscan